Fix UI getting stuck in a scrolled state
Prior to this commit, the UI prevented all scroll attempts if everything
could fit on screen at once. Since the UI isn't usually scrolled when
this happens, this is usually fine. However, removing lines of text can
result in the UI being scrolled while it could fit on screen.

This commit removes that check and allows the regular scrolling logic to
prevent scrolling if necessary.