Ratchet scroll position instead of jumping
When the viewport is scrolled past the end, we want to scroll back up to
an in-bounds position. Previously, this was accomplished with a sudden
"jump" back to the an in-bounds scroll position, which is a bit jarring.
After this commit, the scroll position is instead "ratcheted"; that is,
the viewport can be scrolled up but not down.